+++ src/app/components/SignIn.jsx
@@ -1,6 +1,6 @@
-'use client';
+'use client'
 import { signIn } from 'next-auth/react';
-import Button from '@mui/joy/Button';
+import Button from '@mui/material/Button';
 
 export default function SignIn() {
   return (
@@ -10,7 +10,7 @@ export default function SignIn() {
         href={`/api/auth/signin`}
         onClick={(e) => {
           e.preventDefault();
-          signIn('email');
+          signIn('google');
         }}
       >
         Sign in
