+++ src/app/components/purpleTheme.jsx
@@ -1,0 +1,9 @@
+'use client'
+import { createTheme } from '@mui/material';
+import { purple } from '@mui/material/colors';
+
+export const purpleTheme = createTheme({
+  palette: {
+    primary: purple,
+  },
+});
