+++ src/app/layout.jsx
@@ -1,13 +1,11 @@
-'use client';
 import './globals.css';
 import { Inter } from 'next/font/google';
 import Paper from '@mui/material/Paper';
-import Box from '@mui/material/Box';
-import Button from '@mui/material/Button';
-import Link from 'next/link';
-import { usePathname } from 'next/navigation';
-// import Header from './components/Header';
+import Header from './components/Header';
 import { Provider } from './components/provider';
+import { ThemeProvider } from '@mui/material/styles';
+import BottomNavigation from './components/BottomNavigation';
+import { purpleTheme } from './components/purpleTheme';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -17,35 +15,26 @@ export const metadata = {
 };
 
 export default function RootLayout({ children }) {
-  const pathname = usePathname();
   return (
     <html lang="en">
-      <body className={inter.className}>
-        <Provider>
-          {/* <Header /> */}
-          <main className="flex min-h-screen justify-center py-4 sm:py-24 gap-12">
-            <Paper
-              variant="outlined"
-              color="neutral"
-              className="flex flex-col items-center justify-start rounded-3xl shadow-lg px-2 py-16 gap-8 bg-gradient-to-b from-gray-50 to-slate-50 w-11/12 sm:p-16 lg:w-3/4"
-              sx={{ p: 4 }}
-            >
-              {children}
-              <Box className="self-end mt-auto m-4">
-                {pathname === '/privacy' ? (
-                  <Link href="/">
-                    <Button>Close</Button>
-                  </Link>
-                ) : (
-                  <Link href="privacy">
-                    <Button>Privacy</Button>
-                  </Link>
-                )}
-              </Box>
-            </Paper>
-          </main>
-        </Provider>
-      </body>
+      <ThemeProvider theme={purpleTheme}>
+        <body className={inter.className}>
+          <Provider>
+            {/* <Header /> */}
+            <div className="flex min-h-screen justify-center py-4 sm:py-24 gap-12">
+              <Paper
+                variant="outlined"
+                color="neutral"
+                className="flex flex-col items-center justify-start rounded-3xl shadow-lg px-2 gap-8 bg-gradient-radial from-white to-purple-50 w-11/12 py-0 lg:w-3/4"
+              >
+                <Header />
+                <main className="flex flex-col items-center justify-start gap-8 sm:px-16">{children}</main>
+                <BottomNavigation />
+              </Paper>
+            </div>
+          </Provider>
+        </body>
+      </ThemeProvider>
     </html>
   );
 }
