+++ src/app/privacy/page.jsx
@@ -1,5 +1,5 @@
 export default function Privacy() {
   return (
-        <iframe src="./privacy.html" className="w-full h-full bg-slate-200 rounded-xl p-2" frameborder="0"></iframe>
+        <iframe src="./privacy.html" className="w-full h-full bg-purple-50 shadow-inner rounded-xl p-2" frameborder="0"></iframe>
   );
 }
